Modernizes UI and uses system ffmpeg
Updates the user interface to use Bootstrap 5 for a more modern look and improved responsiveness.

Switches to using system-installed ffmpeg and ffprobe binaries instead of bundled executables. This simplifies deployment, reduces the application size, and aligns with STB-Proxy's approach.  The application now validates the availability of ffmpeg and ffprobe at startup.

Updates the default host address and port.
Addresses an issue in channel selection where the correct host wasn't passed.